+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - DataTables 1.12.1
 - Chart.js v3.9.1
 - chartjs-plugin-datalabels v2.1.0
+- Simplify "New record" notification maintenance #303
 
 ### Fixed
 - handle thousand separator comma
@@ -19,7 +20,7 @@
 ## 4.4.0 - 2022-09-15
 ### Added
 - Navigation drag & drop #292
-- Textvariables for thresholds #299
+- Text variables for thresholds #299
 - New threshold notification type: "new record" #297
 - NC25 support